Break out metric formatter
