+++ datadogclient/datadog_client.go
@@ -3,7 +3,6 @@ package datadogclient
 import (
 	"bytes"
 	"crypto/sha1"
-	"encoding/json"
 	"fmt"
 	"net/http"
 	"sort"
@@ -32,6 +31,7 @@ type Client struct {
 	httpClient            *http.Client
 	maxPostBytes          uint32
 	log                   *gosteno.Logger
+	formatter             Formatter
 }
 
 type MetricKey struct {
@@ -92,6 +92,7 @@ func New(
 		tagsHash:     hashTags(ourTags),
 		httpClient:   httpClient,
 		maxPostBytes: maxPostBytes,
+		formatter:    Formatter{},
 	}
 }
 
@@ -129,19 +130,20 @@ func (c *Client) PostMetrics() error {
 	numMetrics := len(c.metricPoints)
 	c.log.Infof("Posting %d metrics", numMetrics)
 
-	return c.postMetrics(c.metricPoints)
-}
+	c.totalMetricsSent += uint64(len(c.metricPoints))
+	seriesBytes := c.formatter.Format(c.prefix, c.maxPostBytes, c.metricPoints)
+	c.metricPoints = make(map[MetricKey]MetricValue)
 
-func (c *Client) postMetrics(metrics map[MetricKey]MetricValue) error {
-	seriesBytes, metricsCount := formatMetrics(c.prefix, metrics)
-	if uint32(len(seriesBytes)) > c.maxPostBytes {
-		metricsA, metricsB := splitMetrics(metrics)
-		if err := c.postMetrics(metricsA); err != nil {
+	for _, data := range seriesBytes {
+		if err := c.postMetrics(data); err != nil {
 			return err
 		}
-		return c.postMetrics(metricsB)
 	}
 
+	return nil
+}
+
+func (c *Client) postMetrics(seriesBytes []byte) error {
 	url := c.seriesURL()
 	req, err := http.NewRequest("POST", url, bytes.NewBuffer(seriesBytes))
 	req.Header.Set("Content-Type", "application/json")
@@ -158,9 +160,6 @@ func (c *Client) postMetrics(metrics map[MetricKey]MetricValue) error {
 		}
 		return fmt.Errorf("datadog request returned HTTP response: %s\nResponse Body: %s", resp.Status, body)
 	}
-
-	c.totalMetricsSent += metricsCount
-	c.metricPoints = make(map[MetricKey]MetricValue)
 
 	return nil
 }
@@ -208,38 +207,6 @@ func (c *Client) addInternalMetric(name string, value uint64) {
 	}
 
 	c.metricPoints[key] = mValue
-}
-
-func formatMetrics(prefix string, data map[MetricKey]MetricValue) ([]byte, uint64) {
-	metrics := []Metric{}
-	for key, mVal := range data {
-		metrics = append(metrics, Metric{
-			Metric: prefix + key.Name,
-			Points: mVal.Points,
-			Type:   "gauge",
-			Tags:   mVal.Tags,
-		})
-	}
-
-	encodedMetric, _ := json.Marshal(Payload{Series: metrics})
-	return encodedMetric, uint64(len(metrics))
-}
-
-func splitMetrics(data map[MetricKey]MetricValue) (a, b map[MetricKey]MetricValue) {
-	a = make(map[MetricKey]MetricValue)
-	b = make(map[MetricKey]MetricValue)
-	for k, v := range data {
-		split := len(v.Points) / 2
-		a[k] = MetricValue{
-			Tags:   v.Tags,
-			Points: v.Points[:split],
-		}
-		b[k] = MetricValue{
-			Tags:   v.Tags,
-			Points: v.Points[split:],
-		}
-	}
-	return a, b
 }
 
 func getName(envelope *events.Envelope) string {
